requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-ckeditor==6.0.0
django-easymode==0.14.5
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-rosetta==0.9.6
idna==2.10
mysqlclient==2.0.3
Pillow==8.2.0
polib==1.1.1
python-gettext==4.0
pytz==2019.1
requests==2.25.1
six==1.16.0
sqlparse==0.3.0
typing-extensions==3.7.4.3
urllib3==1.26.4
